+++ src/prefabs/goblin.js
@@ -22,7 +22,7 @@ import {
 import { createHumanoidBody } from "./bodies/createHumanoidBody";
 
 import { addSprite } from "../lib/canvas";
-import { updatePosition } from "../lib/ecsHelpers";
+import { updatePosition } from "../ecs/ecsHelpers";
 
 export const createGoblin = (world, options) => {
   const { x, y, z } = options;
