+++ src/ui/menuCharacter.js
@@ -1,7 +1,7 @@
 import _ from "lodash";
 import wrapAnsi from "wrap-ansi";
 import { printRow } from "../lib/canvas";
-import { getEntityData, getWielders } from "../lib/ecsHelpers";
+import { getEntityData, getWielders } from "../ecs/ecsHelpers";
 
 export const renderMenuCharacter = (world, pcEid) => {
   const data = getEntityData(world, pcEid);
